separate the book input component
